+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Store activities for private objects [#3328](https://github.com/opendatateam/udata/pull/3328)
 - Do not crash if file doesn't exists during resource deletion [#3323](https://github.com/opendatateam/udata/pull/3323)
 - Show user domain in suggest [#3324](https://github.com/opendatateam/udata/pull/3324)
+- Keep the existing frequency if not found during harvesting [#3330](https://github.com/opendatateam/udata/pull/3330)
 
 ## 10.4.1 (2025-05-20)