+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - :warning: Add migration to clean duplicate activities [#3327](https://github.com/opendatateam/udata/pull/3327)
 - Do not crash if file doesn't exists during resource deletion [#3323](https://github.com/opendatateam/udata/pull/3323)
 - Show user domain in suggest [#3324](https://github.com/opendatateam/udata/pull/3324)
+- Add new global site metrics [#3325](https://github.com/opendatateam/udata/pull/3325)
 - Keep the existing frequency if not found during harvesting [#3330](https://github.com/opendatateam/udata/pull/3330)
 
 ## 10.4.1 (2025-05-20)