+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Show user domain in suggest [#3324](https://github.com/opendatateam/udata/pull/3324)
 - Add new global site metrics [#3325](https://github.com/opendatateam/udata/pull/3325)
 - Keep the existing frequency if not found during harvesting [#3330](https://github.com/opendatateam/udata/pull/3330)
+- Migrate atom feeds from udata-front to udata [#3326](https://github.com/opendatateam/udata/pull/3326)
 
 ## 10.4.1 (2025-05-20)